requirements.txt
viennarna
dnachisel
redis
celery
Biopython<1.82 # Due to missing CodonUsage module
proglog
docopt
flametree
python_codon_tables
pydantic
numpy < 2
scipy
loguru
click
